refactor(Application.java, BaseBall.java): 야구 클래스 생성
목적에 맞게 야구 게임만을 동작하는 클래스로 리펙토링
